Dodanie powiadomień - aktywacja, dezaktywacja, usuwanie

+++ www/forms/delete.php
@@ -2,13 +2,24 @@
 
 require_once '../config.php';
 
+
 if(isset($_POST['deleteButton']))
 {
+	if ("SELECT FROM users WHERE login = '$delete'")
+	{
+	
     $delete = $_POST['delete'];
     $query = "DELETE FROM users WHERE login = '$delete'";
     $stmt = $db->query($query);
+	
+	header('Location: ../settings_admin.php?not=1');
+	}
+}
+else
+{
+	header('Location: ../settings_admin.php?error=5');
 }
 
-header('Location: ../settings_admin.php');
+
 
 ?>
